% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_project_dependencies.r
\name{install_project_dependencies}
\alias{install_project_dependencies}
\title{Install suggested packages only if missing}
\usage{
install_project_dependencies(
  only_if_missing = TRUE,
  dependencies = c("Imports", "Suggests", "Depends"),
  ...
)
}
\arguments{
\item{only_if_missing}{logical.
If \code{TRUE}, then only installs packages that
are not already installed.
Default is \code{FALSE}.}

\item{dependencies}{character vector.
Types of dependencies to install (e.g., "Imports", "Suggests", "Depends").
Default is \code{c("Imports", "Suggests", "Depends")}.}

\item{...}{arguments passed to \code{install.packages}.}
}
\value{
Invisibly returns character vector of installed packages.
}
\description{
Reads DESCRIPTION file and installs any
suggested packages that are missing.
Useful in R data packages, as this can be run in
a script to create the data to easily install packages
required for actually creating the data, but not required
for simply using the R package as an R package.

Searches for DESCRIPTION file in working directory
and parent of working directory (useful for Rmd's).
}
\details{
Still need/unclear how to do the following:
- install packages specifically off BioConductor (probably
just check if packages are on CRAN first).
- allow installation from remotes.
- allow update of packages
Should we rename function to install_project_dependencies?
- This is not meant for R packages, but rather packages using
`bookdown`, and others where you don't really install the output.
}
