% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_slice.R
\name{view_slice}
\alias{view_slice}
\title{Print slices per group of a dataframe}
\usage{
view_slice(
  data,
  group,
  select = NULL,
  n_print = 1000,
  n_slice = 1,
  seed = NULL,
  sample_within_group = FALSE,
  sample = FALSE,
  prop_sample = NULL,
  return = FALSE
)
}
\arguments{
\item{data}{dataframe. Contains data to print.}

\item{group}{character. Name of column to group by
before subsetting.}

\item{select}{character vector. Name(s) of columns to
select before printing.}

\item{n_print}{numeric. Number of slices to print per group.}

\item{n_slice}{numeric. Number of slices to select per group.}

\item{seed}{numeric. If not \code{NULL}, then seed is
set to this value. Default is \code{NULL}.}

\item{sample_within_group}{logical. If \code{TRUE}
and sampling is selected, then sampling is performed
within each group. Otherwise sampling is performed across
the whole table. Default is \code{FALSE}.}

\item{sample}{logical. If \code{TRUE}, then
rows are sampled. Unless \code{prop_sample} is specified,
then \code{n_slice} rows are sampled.}

\item{prop_sample}{numeric. If not \code{NULL}, then
proportion of rows to sample. Overrides \code{n_slice}.
Default is \code{NULL}.}

\item{return}{logical. If \code{TRUE}, then the subsetted
table is returned. Otherwise, \code{invisible(TRUE)} is
returned. Default is \code{FALSE}.}
}
\description{
Wrapper around

Returns subsetted data invisibly.
}
\examples{
data("cars")
cars[, "grp"] <- rep(letters[1:5], each = 10)
view_slice(cars, group = "grp", n_slice = 2)
}
